New ve Past Challenge siliniyor. Sadece ChallengeFragment kullanılacak

+++ app/src/main/java/com/example/betuldemirci/gout/Model.java
@@ -2,19 +2,22 @@ package com.example.betuldemirci.gout;
 
 public class Model {
 
-    public static final int STACK_TYPE=0;
-    public static final int RUNNING_TYPE=1;
-    public static final int WATER_TYPE=2;
-    public static final int WEIGHT_TYPE=3;
+    public static final int STACK_TYPE = 0;
+    public static final int RUNNING_TYPE = 1;
+    public static final int WATER_TYPE = 2;
+    public static final int WEIGHT_TYPE = 3;
 
     public int type;
-    public int dailyStep, goalStep, calori; // Stack Type
-    public int typeNumber;public String typeName;// 27 mins Running Type
-    public int dailyWater, goalWater;public String waterType; // 5 bottles Water Type
-    public int imageWeight; public double weight; // 80 kg Weight Type
+    public double dailyStep, goalStep, calori; // Stack Type
+    public int typeNumber;
+    public String typeName;// 27 mins Running Type
+    public int dailyWater, goalWater;
+    public String waterType; // 5 bottles Water Type
+    public int imageWeight;
+    public double weight; // 80 kg Weight Type
 
 
-    public Model(int type, int dailyStep, int goalStep, int calori)// Stack Type
+    public Model(int type, double dailyStep, double goalStep, double calori)// Stack Type
     {
         this.type = type;
         this.dailyStep = dailyStep;
@@ -30,7 +33,7 @@ public class Model {
     }
 
     public Model(int type, int dailyWater, int goalWater, String waterType)// Water Type
-        {
+    {
         this.type = type;
         this.dailyWater = dailyWater;
         this.goalWater = goalWater;
